+++ src/routes/(app)/users/[id]/+page.svelte
@@ -58,13 +58,17 @@
             {/await}   
             </span>   
         </div>
+        {user_id}
+        {id}
         {#await req_user_followers(id) then {followers}}
+            {#if user_id != id}
                 {#if followers.includes(user_id)}
                     <button class="Button active" on:click={async () => { await delete_user_unfollows_user(user_id, id) }}>Deixar de seguir</button>
                 {:else}
                     <button class="Button" on:click={async () => { await post_user_follows_user(user_id, id) }}>Seguir</button>
                 {/if}
-            {/await}
+            {/if}
+        {/await}
     </div>
     <div class="gap-4 flex flex-col items-left last-matches" style="flex-grow: 1;">
         <h1>Reviews do usuário:</h1>
